requirements.txt
langchain
langchain-community
numpy
pandas
streamlit
langchain-groq
python-dotenv
chromadb
tiktoken
sentence-transformers
bs4
langchainhub
wikipedia
langchain-core
firestore
flask
